backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
celery
redis
psycopg2-binary
django-cors-headers
drf-spectacular
pytest
pytest-django
pytest-cov
